ERA-10338: Add i18n and refacto the sound notifications player to separate concerns per type of notification
