+++ public/locales/en-US/menu-drawer.json
@@ -5,8 +5,8 @@
   },
   "betaToggles": {
     "playSoundForAlarmRadioState": "Play sound and show toast when a radio state changes to red (alarm)",
-    "playSoundForNewInReachMessage": "Play Sound For New InReach Messages",
-    "reportNotification": "Play Sound For New Events"
+    "playSoundForNewInReachMessage": "Play sound for new inReach messages",
+    "reportNotification": "Play sound for new events"
   },
   "dailyReportModal": {
     "dataExportModalTitle": "Daily Report",
