+++ public/locales/es/menu-drawer.json
@@ -5,7 +5,7 @@
   },
   "betaToggles": {
     "playSoundForAlarmRadioState": "Reproducir alerta y mostrar mensaje cuando un radio cambie a estado rojo (alerta)",
-    "playSoundForNewInReachMessage": "TODO",
+    "playSoundForNewInReachMessage": "Reproducir alerta para nuevos mensajes de inReach",
     "reportNotification": "Reproducir alerta para nuevos eventos"
   },
   "dailyReportModal": {
