+++ src/SoundNotificationsPlayer/RadioStateChangeToAlarmNotificationPlayer/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../common/styles/vars/colors';
+@import '../../common/styles/vars/colors';
 
 .toast {
   background-color: $bright-red;
